Added List Page
Added List page to program, the list is not able to pull data from firestore but functions are in place to set it up (I think it's mainly my html table that is causing difficulty). Other fixes are minor fixes I made throughout the program to keep it consistent. signup can be ignored but thought I would add in case we go back to that last minute. Added User List to NavBar